chore(rules): relocate caps-spec artifacts to src/rules/ and tests/rules-vectors/ (#715)
Atomic GREEN commit for caps-spec-consolidation PR1:
- git mv all 7 rule artifacts: src/vendor/caps-spec/ -> src/rules/ (flat)
- git mv test vector: src/vendor/caps-spec/test-vectors/ -> tests/rules-vectors/
- repath runtime imports in affiliates.js and wrapper-engine.js
- repath CAPS_DIRECT_INJECTION_PROGRAMS import in caps-manifest-sync.test.mjs
- update .gitattributes eol=lf pins from vendor paths to src/rules/ paths
- all 4027 tests pass; Ed25519 sig verifies from new path; conformance:contextual green
