+++ src/rules/manifest.data.js
