+++ src/rules/manifest.json
